Remove @testable annotations where possible
