+++ Tests/RegexTests/CompileTests.swift
@@ -9,7 +9,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-
 @testable import _RegexParser
 @testable import _StringProcessing
